+++ source/dust/platform/win32/compute_shader.cpp
@@ -7,12 +7,12 @@ using namespace dust;
 using dust::compute_shader_t;
 
 
-auto dust::create_compute_shader(context_ptr const& context, void const* data, uint data_size) -> compute_shader_ptr
+auto dust::create_compute_shader(context_ptr const& context, void const* data, size_t data_size) -> compute_shader_ptr
 {
 	return compute_shader_ptr(new compute_shader_t(context, data, data_size));
 }
 
-compute_shader_t::compute_shader_t(context_ptr const& context, void const* data, uint data_size)
+compute_shader_t::compute_shader_t(context_ptr const& context, void const* data, size_t data_size)
 : context_(context)
 {
 	auto const& device = context_->d3d_device();
